Refine types on apoc procedures
