+++ packages/cypher-builder/src/apoc/functions/RunFirstColumn.ts
