+++ packages/cypher-builder/src/types.ts
@@ -35,8 +35,7 @@ import type { ListExpr } from "./expressions/list/ListExpr";
 import type { MapProjection } from "./expressions/map/MapProjection";
 import type { HasLabel } from "./expressions/HasLabel";
 import type { Reference } from "./variables/Reference";
-import type { Validate } from "./procedures/apoc/Validate";
-import type { ApocExpr, ApocPredicate } from "./procedures/apoc/apoc";
+import type { ApocFunction, ApocPredicate, ApocProcedure } from "./apoc/types";
 
 export type Operation = BooleanOp | ComparisonOp | MathOp;
 
@@ -52,7 +51,7 @@ export type Expr =
     | MapExpr // NOTE this cannot be set as a property in a node
     | MapProjection // NOTE this cannot be set as a property in a node
     | ListExpr
-    | ApocExpr;
+    | ApocFunction;
 
 /** Represents a predicate statement (i.e returns a boolean). Note that RawCypher is only added for compatibility */
 export type Predicate =
@@ -67,7 +66,7 @@ export type Predicate =
     | HasLabel;
 
 /** Represents a procedure, invocable with the CALL statement */
-export type Procedure = Validate;
+export type Procedure = ApocProcedure;
 
 export type CypherResult = {
     cypher: string;
